api/workers/celery-worker/requirements.txt
celery==5.4.0
redis==5.0.8
fastapi==0.115.2
uvicorn==0.30.6
pillow==10.4.0
ImageHash==4.3.1
requests==2.32.3
flower==2.0.1
